+++ src/core/analysis-context.ts
@@ -14,6 +14,7 @@
 
 import * as path from 'path';
 
+
 import {ForkOptions, LazyEdgeMap, NoKnownParserError, Options, ScannerTable} from '../core/analyzer';
 import {CssCustomPropertyScanner} from '../css/css-custom-property-scanner';
 import {CssParser} from '../css/css-parser';